+++ .gitignore
@@ -89,4 +89,4 @@ npm-debug.log
 .DS_Store
 
 uploads/*
-src/repo
+repo/*
