+++ .idea/vcs.xml
@@ -2,6 +2,6 @@
 <project version="4">
   <component name="VcsDirectoryMappings">
     <mapping directory="$PROJECT_DIR$" vcs="Git" />
-    <mapping directory="$PROJECT_DIR$/src/repo" vcs="Git" />
+    <mapping directory="$PROJECT_DIR$/repo" vcs="Git" />
   </component>
 </project>
