+++ src/controller/SpecificImageController.ts
@@ -1,16 +1,16 @@
 import * as express from 'express'
-import {ShowcaseController} from "./ShowcaseController";
+import {ShowcaseFacade} from "../main/ShowcaseFacade";
 import {IControllerBase} from "../model/CustomInterfaces";
 
 export class SpecificImageController implements IControllerBase {
     public path = '/showcase/:uid';
     public router = express.Router();
 
-    constructor(showcaseController: ShowcaseController) {
-        this.initRoutes(showcaseController);
+    constructor(facade: ShowcaseFacade) {
+        this.initRoutes(facade);
     }
 
-    public initRoutes(showcaseController: ShowcaseController) {
-        this.router.delete(this.path, showcaseController.deleteImage);
+    public initRoutes(facade: ShowcaseFacade) {
+        this.router.delete(this.path, facade.deleteImage);
     }
 }
