+++ src/model/CustomErrors.ts
